update-profile:
Add user update support and clarify service responsibilities

- Add UserUpdateRequestDto, identical in structure to RegisterRequestDto, but kept separate for semantic clarity.
- Refactor UserProfileService to use UserRepository directly instead of UserManagementService: UserManagementService will handle business operations on users, while UserProfileService will be kept focused on authenticated user profile actions.

Add the Endpoint to the RestController
- Add to UserController a @PutMapping endpoint for the profile update

Manage the update problem:

Because update method can now update the email or username and the user can authenticate with one or the other, we should manage the update of the security context. Another method is to encode the user id into the jwt claimset. This id never change.

- change method AuthenticationService.getAuthenticatedUser used to retrieve authenticated User. The id is extract from claims and used to get the user from repository instead of using authentication.getName()
- update DefaultJwtTokenProvider to include id claim into jwt token during token generation

Refactor:

Replace AuthenticationService.getAuthenticatedUser and AuthenticationService.getAuthenticatedUserEmail by the new AuthenticationService.getAuthenticatedUser
=> simplification in services by replacing (getAuthenticatedUserEmail + getAuthenticatedUser with UsermanagementService)

clean some comments, debug lines and linting